backend/requirements.txt
strands-agents
bedrock-agentcore
boto3
fastapi
uvicorn[standard]
python-dotenv
pydantic
aiofiles
httpx
twilio
elevenlabs
websockets
pyaudio
aws-sdk-bedrock-runtime
smithy-aws-core
awscrt
